BACKLOG-2133 - Datastore migration : Integrate job in the fix-applier or in a dedicated tool http://jira.jahia.org/browse/BACKLOG-2133 https://devtools.jahia.com/fisheye/cru/CJ-HOT-104
The migration job is implemented as part of Jahia core and not as part of the fix applier as the assigned time was not enough to investigate the problem with composing correct classpath and parsing database parameters from the context.xml to create DB connection.

How the migration tool works:
1) On Jahia startup before starting the Jackrabbit repository it is checked if the repository has no DataStore. If not the migration is started.
2) Temporary DB tables are created for persistence managers (default/live/version, global and datastore).
3) A copy of repository.xml is created and adjusted to use datastore and temporary tables
4) A RepositoryCopier is executed to copy the source repository to target using the created configuration
5) Original repository is deleted (on demand backuped if the corresponding system flag is set to true)
6) The original Jackrabbit tables (except several, not related) are deleted
7) The temporary tables (with the migrated data) are rename including indexes
8) The repository.xml is adjusted to point to those renamed tables
9) The Lucene indexes of the migrated repository are deleted to trigger indexation
10) The Jahia server startup continues with the repository startup (the indexation is done here)

I had to correct the SpellChecker to properly close IndexReader when the repository is shut down (see JahiaExtendedSpellChecker and CompositeSpellChecker changes).

Additionally, I've added a Groovy migration script to trigger spellchecker index rebuild after the server startup

TODO: scripts for Oracle and MS SQL Server

git-svn-id: https://devtools.jahia.com/svn/jahia_pack/trunk@232 39555585-ee67-472c-bf1e-947fb1f6b7c6