BACKLOG-2253 - RB improvements http://jira.jahia.org/browse/BACKLOG-2253
Make native2ascii task easier to configure (move the most part to the parent jahia-modules).
Added support for creating the default-locale resource bundle if not present (the locale of the default bundle is configurable and is set by default to "en").

TODO: adjust the jahia-modules pom.xml and sub-module configuration

git-svn-id: https://devtools.jahia.com/svn/jahia_pack/branches/JAHIA-6-6-X-X-BRANCH@261 39555585-ee67-472c-bf1e-947fb1f6b7c6